--- components/Common/DarkModeToggle/useTheme.tsx
@@ -1,6 +1,0 @@
-import { useContext } from 'react';
-import { ThemeContext } from '../../../providers/themeprovider';
-
-export function useTheme() {
-  return useContext(ThemeContext) ?? [];
-}
