+++ package.json
@@ -47,6 +47,7 @@
     "isomorphic-dompurify": "^1.1.0",
     "next": "13.2.0",
     "next-sitemap": "^4.0.6",
+    "next-themes": "^0.2.1",
     "nextra": "^2.2.20",
     "node-version-data": "^1.1.0",
     "npm-run-all": "^4.1.5",
